feat: Add UserMenu component and integrate it into the dashboard layout
refactor: Simplify AppSidebar by removing unused code and integrating MyProject and ParticipatedProjects components

feat: Implement ProjectCardGrid and ProjectCardSkeleton for project display

chore: Remove deprecated ProjectOverview component and related logic

fix: Update edit-task-modal to handle optional fields correctly

feat: Add middleware for authentication and redirect logic

chore: Update store configuration to remove unnecessary function wrapper

style: Adjust tsconfig to include middleware for type checking #15
