+++ src/components/dashboard/UserMenu.tsx
@@ -1,0 +1,59 @@
+'use client';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from '../ui/dropdown-menu';
+import { Avatar, AvatarFallback, AvatarImage } from '../ui/avatar';
+import { logoutUserAPI, selectCurrentUser } from '@/lib/features/auth/authSlice';
+import { useAppDispatch } from '@/lib/hooks';
+import { useSelector } from 'react-redux';
+import { Button } from '../ui/button';
+import Link from 'next/dist/client/link';
+
+import { Settings, User as UserIcon, LogOut } from 'lucide-react';
+
+export function UserMenu() {
+  const dispatch = useAppDispatch();
+  const currentUser = useSelector(selectCurrentUser);
+  return (
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button variant="ghost" className="gap-2 h-auto py-2 px-3">
+              <Avatar className="w-8 h-8">
+                <AvatarImage src={currentUser?.avatar} />
+                <AvatarFallback className="bg-gradient-to-br from-blue-500 to-purple-600 text-white">
+                  {currentUser?.displayName.substring(0, 2).toUpperCase()}
+                </AvatarFallback>
+              </Avatar>
+              <span className="text-sm text-gray-700">{currentUser?.displayName}</span>
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent align="end" className="w-56">
+            <DropdownMenuLabel>My Account</DropdownMenuLabel>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem asChild>
+              <Link href="/profile">
+              <UserIcon className="w-4 h-4 mr-2" />
+                <span>Profile</span>
+              </Link>
+            </DropdownMenuItem>
+            <DropdownMenuItem>
+              <Settings className="w-4 h-4 mr-2" />
+              <span>Settings</span>
+            </DropdownMenuItem>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem 
+              className="text-red-600"
+              onClick={() => dispatch(logoutUserAPI())}  
+            >
+              <LogOut className="w-4 h-4 mr-2" />
+              <span>Log out</span>
+            </DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+  )
+}
