+++ src/lib/hooks/useProjects.ts
@@ -1,3 +1,4 @@
+'use client'
 import {
   useQuery,
   useMutation,
