+++ tsconfig.json
@@ -4,7 +4,7 @@
     "next-env.d.ts",
     "src/**/*",
     ".next/types/**/*.ts"
-  ],
+, "src/middleware.ts"  ],
   "exclude": [
     "node_modules",
     ".next"
